[reorganization]:Move inheritance logic to its own app
Fixes #7334
